Rust

+++ src/utils/color.rs
@@ -17,10 +17,10 @@ impl Default for Color {
 }
 
 impl Color {
-    pub fn rgba8888to_color(&mut self, value: u32) {
-        self.r = (((value & 0xff000000) >> 24) / 255) as f32;
-        self.g = (((value & 0x00ff0000) >> 16) / 255) as f32;
-        self.b = (((value & 0x0000ff00) >> 8) / 255) as f32;
-        self.a = ((value & 0x000000ff) / 255) as f32;
+    pub fn rgba8888to_color(&mut self, value: i32) {
+        self.r = (((value as u32 & 0xff000000) >> 24) / 255) as f32;
+        self.g = (((value as u32 & 0x00ff0000) >> 16) / 255) as f32;
+        self.b = (((value as u32 & 0x0000ff00) >> 8) / 255) as f32;
+        self.a = ((value as u32 & 0x000000ff) / 255) as f32;
     }
 }
